move menu to camera settings
